# Conflicts:
#	connectiontypes/pom.xml
#	dlms/pom.xml
#	dlms/src/main/java/com/energyict/dlms/DlmsSession.java
#	dlms/src/main/java/com/energyict/dlms/protocolimplv2/ApplicationServiceObjectV2.java
#	dlms/src/main/java/com/energyict/dlms/protocolimplv2/DlmsSession.java
#	dlms/src/main/java/com/energyict/dlms/protocolimplv2/connection/SecureConnection.java
#	dlms/src/main/java/com/energyict/encryption/asymetric/util/KeyUtils.java
#	pom.xml
#	protocolcommon/pom.xml
#	protocolcommon/src/main/java/com/energyict/protocolimplv2/messages/SecurityMessage.java
#	protocolimpl-elster/pom.xml
#	protocolimpl-us/pom.xml
#	protocolimpl-v2/pom.xml
#	protocolimpl-v2/src/main/java/com/energyict/mdc/protocol/inbound/g3/EventPushNotificationParser.java
#	protocolimpl-v2/src/main/java/com/energyict/mdc/protocol/inbound/g3/PushEventNotification.java
#	protocolimpl-v2/src/main/java/com/energyict/mdc/protocol/inbound/idis/DataPushNotification.java
#	protocolimpl-v2/src/main/java/com/energyict/mdc/protocol/inbound/idis/DataPushNotificationParser.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/dlms/g3/AS330D.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/dlms/idis/am130/AM130.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/dlms/idis/am540/AM540.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtu3/beacon3100/Beacon3100.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtu3/beacon3100/messages/Beacon3100Messaging.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtu3/beacon3100/properties/Beacon3100ConfigurationSupport.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtu3/beacon3100/properties/Beacon3100Properties.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtu3/beacon3100/properties/Beacon3100SecurityProvider.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/rtuplusserver/g3/RtuPlusServer.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/nta/abstractnta/NTASecurityProvider.java
#	protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/nta/dsmr23/DlmsProperties.java
#	protocolimpl/pom.xml

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/eict/eiweb/PacketBuilder.java
@@ -257,9 +257,9 @@ public class PacketBuilder {
         }
     }
 
-    private int[] parseValues(byte[] buffer) {
+    private long[] parseValues(byte[] buffer) {
         String strValues = new String(buffer);
-        int[] valuesArray = new int[nrOfChannels];
+        long[] valuesArray = new long[nrOfChannels];
         StringTokenizer st = new StringTokenizer(strValues, ",");
         int iTokens = st.countTokens();
         if (iTokens != nrOfChannels) {
@@ -267,7 +267,7 @@ public class PacketBuilder {
         }
         try {
             for (int i = 0; i < iTokens; i++) {
-                valuesArray[i] = Integer.parseInt(st.nextToken());
+                valuesArray[i] = Long.parseLong(st.nextToken());
             }
         } catch (NumberFormatException e) {
             throw CommunicationException.numberFormatException(e, EIWebConstants.METER_DATA_PARAMETER_NAME, strValues);
@@ -275,7 +275,7 @@ public class PacketBuilder {
         return valuesArray;
     }
 
-    private void createData(String utc, String code, String statebits, int[] values) throws IOException {
+    private void createData(String utc, String code, String statebits, long[] values) throws IOException {
         ByteArrayOutputStream bos = new ByteArrayOutputStream();
         LittleEndianOutputStream os = new LittleEndianOutputStream(bos);
         this.addUTCToDataStream(utc, os);
@@ -311,10 +311,10 @@ public class PacketBuilder {
         }
     }
 
-    private void addValuesToDataStream(int[] values, LittleEndianOutputStream os) throws IOException {
+    private void addValuesToDataStream(long[] values, LittleEndianOutputStream os) throws IOException {
         for (int i = 0; i < values.length; i++) {
             if (version < VERSION_32BITS_3) {
-                os.writeLEShort((short) values[i]);
+                os.writeLEShort(values[i]);
             } else {
                 os.writeLEInt(values[i]);
             }
